+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased 3.0](https://github.com/opensearch-project/k-NN/compare/2.x...HEAD)
 ### Features
 * upgrade to opensearch 3.0.0-SNAPSHOT and remove Java SecurityManager
+* Add jVector search query statistics [#62](https://github.com/opensearch-project/opensearch-jvector/issues/62)
 ### Enhancements
 * Upgrade to java 22 so that we can use Foreign Memory API and MemorySegmentReader
 ### Bug Fixes